Leave infix applications alone, forever.
Infix applications have caused way too much trouble that I've given up
entirely on trying to come up with anything reasonable for them. This
commit changes scalafmt so that it will keep line breaking decisions in
all infix applications. The rhs of the application is indented by 2
spaces if the application is not a "top-level" statement, per Scala.js
coding style:

https://github.com/scala-js/scala-js/blob/master/CODINGSTYLE.md#long-expressions-with-binary-operators

* Fixes #269
* Fixes #267
* Fixes #284